python/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
transformers==4.38.2
openai-whisper==20231117
torch==2.2.1
sentence-transformers==2.6.1
newspaper3k==0.2.8
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
serpapi==0.1.5
mediapipe==0.10.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2